Quick & dirty shrink operand size to 16-bits

+++ Zend/Optimizer/block_pass.c
@@ -1148,8 +1148,8 @@ static void assemble_code_blocks(zend_cfg *cfg, zend_op_array *op_array, zend_op
 				zend_op *end = opline + len;
 				while (opline < end) {
 					if (opline->opcode == ZEND_FAST_RET &&
-					    opline->op2.num != (uint32_t)-1 &&
-					    opline->op2.num < (uint32_t)j) {
+					    opline->op2.num != (uint16_t)-1 &&
+					    opline->op2.num < (uint16_t)j) {
 						opline->op2.num = map[opline->op2.num];
 					}
 					opline++;
