+++ Zend/zend_compile.h
@@ -34,7 +34,7 @@
 
 #define SET_UNUSED(op) do { \
 	op ## _type = IS_UNUSED; \
-	op.num = (uint32_t) -1; \
+	op.num = (uint16_t) -1; \
 } while (0)
 
 #define MAKE_NOP(opline) do { \
@@ -67,14 +67,14 @@ typedef struct _zend_op zend_op;
 #endif
 
 typedef union _znode_op {
-	uint32_t      constant;
-	uint32_t      var;
-	uint32_t      num;
-	uint32_t      opline_num; /*  Needs to be signed */
+	uint16_t      constant;
+	uint16_t      var;
+	uint16_t      num;
+	uint16_t      opline_num; /*  Needs to be signed */
 #if ZEND_USE_ABS_JMP_ADDR
 	zend_op       *jmp_addr;
 #else
-	uint32_t      jmp_offset;
+	uint16_t      jmp_offset;
 #endif
 #if ZEND_USE_ABS_CONST_ADDR
 	zval          *zv;
@@ -140,7 +140,7 @@ struct _zend_op {
 	znode_op op1;
 	znode_op op2;
 	znode_op result;
-	uint32_t extended_value;
+	uint16_t extended_value;
 	uint32_t lineno;
 	uint8_t opcode;       /* Opcodes defined in Zend/zend_vm_opcodes.h */
 	uint8_t op1_type;     /* IS_UNUSED, IS_CONST, IS_TMP_VAR, IS_VAR, IS_CV */
@@ -163,7 +163,7 @@ typedef struct _zend_slim_op {
 	znode_op op1;
 	znode_op op2;
 	znode_op result;
-	uint32_t extended_value;
+	uint16_t extended_value;
 } zend_slim_op;
 
 typedef struct _zend_brk_cont_element {
@@ -745,8 +745,8 @@ ZEND_STATIC_ASSERT(ZEND_MM_ALIGNED_SIZE(sizeof(zval)) == sizeof(zval),
 #define EX_VAR(n)				ZEND_CALL_VAR(execute_data, n)
 #define EX_VAR_NUM(n)			ZEND_CALL_VAR_NUM(execute_data, n)
 
-#define EX_VAR_TO_NUM(n)		((uint32_t)((n) / sizeof(zval) - ZEND_CALL_FRAME_SLOT))
-#define EX_NUM_TO_VAR(n)		((uint32_t)(((n) + ZEND_CALL_FRAME_SLOT) * sizeof(zval)))
+#define EX_VAR_TO_NUM(n)		((uint16_t)((n) / sizeof(zval) - ZEND_CALL_FRAME_SLOT))
+#define EX_NUM_TO_VAR(n)		((uint16_t)(((n) + ZEND_CALL_FRAME_SLOT) * sizeof(zval)))
 
 #define ZEND_OPLINE_TO_OFFSET(opline, target) \
 	((char*)(target) - (char*)(opline))
@@ -756,10 +756,10 @@ ZEND_STATIC_ASSERT(ZEND_MM_ALIGNED_SIZE(sizeof(zval)) == sizeof(zval),
 
 #define ZEND_OFFSET_TO_OPLINE(base, offset) \
 	_Generic((base), \
-		zend_op*: ((zend_op*)(((char*)(base)) + (int)offset)), \
-		const zend_op*: ((zend_op*)(((char*)(base)) + (int)offset)), \
-		zend_slim_op*: ((zend_slim_op*)(((char*)(base)) + (int)offset)), \
-		const zend_slim_op*: ((zend_slim_op*)(((char*)(base)) + (int)offset)))
+		zend_op*: ((zend_op*)(((char*)(base)) + (int16_t)offset)), \
+		const zend_op*: ((zend_op*)(((char*)(base)) + (int16_t)offset)), \
+		zend_slim_op*: ((zend_slim_op*)(((char*)(base)) + (int16_t)offset)), \
+		const zend_slim_op*: ((zend_slim_op*)(((char*)(base)) + (int16_t)offset)))
 
 #define ZEND_OFFSET_TO_OPLINE_NUM(op_array, base, offset) \
 	(ZEND_OFFSET_TO_OPLINE(base, offset) - op_array->opcodes)
